# src/main/docker/openapi.yaml
{"openapi":"3.0.1","info":{"title":"Application API","description":"appDescription","contact":{"name":"Svyatoslav","email":"test@gmail.com"},"license":{"name":"Apache 2.0","url":"http://springdoc.org"},"version":"appVersion"},"servers":[{"url":"http://localhost:8080","description":"Dev service"},{"url":"http://localhost:8082","description":"Beta service"}],"paths":{"/card":{"get":{"tags":["card-controller"],"operationId":"get","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}}}},"put":{"tags":["card-controller"],"operationId":"putBalance","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransferMoneyDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}}}}},"/auth/register":{"post":{"tags":["auth-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDto"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"auth","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CardDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/card/transaction":{"get":{"tags":["card-controller"],"operationId":"getTransactions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionDto"}}}}}}},"/card/availableOperationType":{"get":{"tags":["card-controller"],"operationId":"availableOperationType","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string","enum":["TRANSFER_BETWEEN_ACCOUNTS","REPLENISHMENT","WITHDRAWAL"]}}}}}}}}},"components":{"schemas":{"TransferMoneyDto":{"type":"object","properties":{"toNumber":{"type":"string"},"operationType":{"type":"string","enum":["TRANSFER_BETWEEN_ACCOUNTS","REPLENISHMENT","WITHDRAWAL"]},"amount":{"type":"number"}}},"CardDto":{"type":"object","properties":{"number":{"type":"string"},"pinCode":{"type":"string"},"balance":{"type":"number"}}},"Data":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"},"interaction":{"type":"string"},"balanceChangeFrom":{"type":"number"},"balanceChangeTo":{"type":"number"}}},"TransactionDto":{"type":"object","properties":{"cardDto":{"$ref":"#/components/schemas/CardDto"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Data"}}}}}}}